added portal and refactored a little
