+++ components/auth/Signin.js
@@ -22,7 +22,7 @@ export default function SignIn({ login }) {
         password: values.password,
       });
       if (!response.error) {
-        router.replace("/");
+        router.replace("/portal");
         resetForm();
       } else {
         setError(response.error);
